+++ vnpy/app/spread_trading/engine.py
@@ -441,7 +441,7 @@ class SpreadAlgoEngine:
         offset: Offset,
         price: float,
         volume: float,
-        payup: int,
+        payup: float,
         interval: int,
         lock: bool,
         spread_rate: float
@@ -926,7 +926,7 @@ class SpreadStrategyEngine:
         offset: Offset,
         price: float,
         volume: float,
-        payup: int,
+        payup: float,
         interval: int,
         spread_rate: float,
         lock: bool,
